removed selector file out of store
