+++ src/helpers/utils.js
@@ -1,3 +1,36 @@
+export const excludeCurrentOffer = (offers, currentOfferId) => offers.filter((item) => item.id !== currentOfferId);
+
+export const reduceOffers = (offers) => offers.length <= 3 ? offers : offers.slice(0, 10);
+
+const sortReveiwsByDate = (reviews) => (reviews.sort((a, b) => (new Date(b.date) - new Date(a.date))));
+
+const reduceReviews = (reviews) => reviews.length <= 10 ? reviews : reviews.slice(0, 10);
+
+export const getReviews = (reviews) => reduceReviews(sortReveiwsByDate(reviews));
+
+export const reduceCities = (cities) => cities.length >= 6 ? cities : cities.slice(0, 6);
+
+export const sortByPrice = (offers) => (offers.sort((a, b) => b.price - a.price));
+
+export const sortByPriceReverse = (offers) => sortByPrice(offers).reverse();
+
+export const sortByRating = (offers) => (offers.sort((a, b) => b.rating - a.rating));
+
+export const filterOffersOrder = (offers, filter) => {
+  switch (filter) {
+    case FILTERS_MAP.popular:
+      return offers;
+    case FILTERS_MAP.priceLowToHight:
+      return sortByPrice(offers);
+    case FILTERS_MAP.priceHightToLow:
+      return sortByPriceReverse(offers);
+    case FILTERS_MAP.topRatedFirst:
+      return sortByRating(offers);
+    default:
+      return offers;
+  }
+};
+
 export const extend = (a, b) => Object.assign({}, a, b);
 
 export const getRatingInPercents = (rating) => `${(Math.round(rating)) * 20}%`;
@@ -80,4 +113,3 @@ export const formatReview = (review) => ({
 });
 
 export const formatReviews = (reviews) => reviews.map((review) => formatReview(review));
-
